# openapi3_1.yaml
openapi: '3.1.0'
info:
  version: '1.0.0'
  title: 'Food Driver API'
  description: System obsługi zamówień online dla restauracji
servers:
  - url: http://localhost:3000/api
    description: Środowisko deweloperskie
paths:
...
  /users/me/orders:
    get:
      operationId: ListMeOrders
      description: Przegląd zamówień zalogowanego użytkownika
      security:
        - bearerAuth: [Klient]
      parameters:
        - name: status
          description: Status zamówienia
          in: query
          schema:
            $ref: '#/components/schemas/OrderStatus'
          required: false
        - name: reported
          description: Zgłoszono problem z zamówieniem
          in: query
          schema:
            type: boolean
          required: false
        - name: page
          description: Numer strony
          in: query
          schema:
            $ref: '#/components/schemas/Page'
          required: false
      responses:
        '200':
          description: Przegląd zamówień użytkownika
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                    minimum: 0
                  pages:
                    type: integer
                    minimum: 1
                  currentPage:
                    type: integer
                    minimum: 1
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        idOrder:
                          $ref: '#/components/schemas/Id'
                        restaurant:
                          type: object
                          properties:
                            idRestaurant:
                              $ref: '#/components/schemas/Id'
                            name:
                              $ref: '#/components/schemas/Name'
                            category:
                              type: object
                              properties:
                                idCategory:
                                  $ref: '#/components/schemas/Id'
                                name:
                                  $ref: '#/components/schemas/Name'
                              required:
                                - idCategory
                                - name
                              additionalProperties: false
                          required:
                            - idRestaurant
                            - name
                            - category
                          additionalProperties: false
                        status:
                          $ref: '#/components/schemas/OrderStatus'
                        payment:
                          $ref: '#/components/schemas/Price'
                        reported:
                          type: boolean
                        created:
                          $ref: '#/components/schemas/DateTime'
                      required:
                        - idOrder
                        - restaurant
                        - status
                        - payment
                        - reported
                        - created
                      additionalProperties: false
                required:
                  - count
                  - pages
                  - currentPage
                  - results
                additionalProperties: false
        '401':
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/InternalServerError'
      tags:
        - Users
...
  /users/me/restaurant/orders:
    get:
      operationId: ListMeRestaurantOrders
      description: Przegląd zamówień restauracji
      security:
        - bearerAuth: [Restaurator]
      parameters:
        - name: status
          description: Status zamówienia
          in: query
          schema:
            $ref: '#/components/schemas/OrderStatus'
          required: false
        - name: reported
          description: Zgłoszono problem z zamówieniem
          in: query
          schema:
            type: boolean
          required: false
        - name: page
          description: Numer strony
          in: query
          schema:
            $ref: '#/components/schemas/Page'
          required: false
      responses:
        '200':
          description: Przegląd zamówień restauracji
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                    minimum: 0
                  pages:
                    type: integer
                    minimum: 1
                  currentPage:
                    type: integer
                    minimum: 1
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        idOrder:
                          $ref: '#/components/schemas/Id'
                        restaurant:
                          type: object
                          properties:
                            idRestaurant:
                              $ref: '#/components/schemas/Id'
                            name:
                              $ref: '#/components/schemas/Name'
                            category:
                              type: object
                              properties:
                                idCategory:
                                  $ref: '#/components/schemas/Id'
                                name:
                                  $ref: '#/components/schemas/Name'
                              required:
                                - idCategory
                                - name
                              additionalProperties: false
                          required:
                            - idRestaurant
                            - name
                            - category
                          additionalProperties: false
                        status:
                          $ref: '#/components/schemas/OrderStatus'
                        payment:
                          $ref: '#/components/schemas/Price'
                        reported:
                          type: boolean
                        created:
                          $ref: '#/components/schemas/DateTime'
                      required:
                        - idOrder
                        - restaurant
                        - status
                        - payment
                        - reported
                        - created
                      additionalProperties: false
                required:
                  - count
                  - pages
                  - currentPage
                  - results
                additionalProperties: false
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        default:
          $ref: '#/components/responses/InternalServerError'
      tags:
        - Users
...
  /orders/{id}/report:
    parameters:
      - name: id
        description: Unikalny identyfikator zamówienia
        in: path
        schema:
          $ref: '#/components/schemas/Id'
        required: true
    post:
      operationId: CreateOrderReport
      description: Utworzenie zgłoszenia problemu z zamówieniem
      security:
        - bearerAuth: [Klient]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  $ref: '#/components/schemas/Description'
                idImage:
                  $ref: '#/components/schemas/Id'
              required:
                - message
              additionalProperties: false
      responses:
        '204':
          description: Pomyślnie utworzono zgłoszenie problemu z zamówieniem
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/InternalServerError'
      tags:
        - Orders
  /orders/{id}/cancel:
    parameters:
      - name: id
        description: Unikalny identyfikator zamówienia
        in: path
        schema:
          $ref: '#/components/schemas/Id'
        required: true
    post:
      operationId: CancelOrder
      description: Anulowanie zamówienia
      security:
        - bearerAuth: [Restaurator, Administrator]
      responses:
        '204':
          description: Pomyślnie anulowano zamówienie
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        default:
          $ref: '#/components/responses/InternalServerError'
      tags:
        - Orders
  /orders/{id}/resign:
    parameters:
      - name: id
        description: Unikalny identyfikator zamówienia
        in: path
        schema:
          $ref: '#/components/schemas/Id'
        required: true
    post:
      operationId: ResignToDeliverOrder
      description: Rezygnacja z dostarczenia zamówienia
      security:
        - bearerAuth: [Dostawca, Administrator]
      responses:
        '204':
          description: Pomyślnie zrezygnowano z dostarczenia zamówienia
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        default:
          $ref: '#/components/responses/InternalServerError'
      tags:
        - Orders
  /orders/{id}/accept-for-realization:
    parameters:
      - name: id
        description: Unikalny identyfikator zamówienia
        in: path
        schema:
          $ref: '#/components/schemas/Id'
        required: true
    post:
      operationId: AcceptOrderForRealization
      description: Przyjęcie zamówienia do realizacji
      security:
        - bearerAuth: [Restaurator]
      responses:
        '204':
          description: Pomyślnie przyjęto zamówienie do realizacji
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        default:
          $ref: '#/components/responses/InternalServerError'
      tags:
        - Orders
...
components:
  responses:
...
    BadRequest:
      description: Nieprawidłowa treść żądania
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Nieprawidłowy token autentykacji
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Użytkownik nie może wykonać tej akcji
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalServerError:
      description: Wystąpił wewnętrzny błąd serwera
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Page:
      description: Numer strony
      type: integer
      minimum: 1
    Id:
      description: Unikalny identyfikator
      type: integer
...
    Name:
      type: string
      minLength: 1
      maxLength: 200
    Description:
      type: string
      minLength: 1
      maxLength: 3000
...
    Price:
      type: string
      format: ^\d+(.\d{1,2})?$
...
    DateTime:
      type: string
      format: date-time
...
    OrderStatus:
      type: string
      enum:
        - Nowe
        - Przyjęte do realizacji
        - Gotowe do dostawy
        - Odebrane do dostarczenia
        - Dostarczone
        - Zrealizowane
        - Anulowane
...
    Error:
      type: object
      properties:
        message:
          description: Komunikat dla użytkownika
          type: string
      required:
        - message
      additionalProperties: false
  securitySchemes:
    bearerAuth:
      description: Token na okaziciela JWT
      type: http
      scheme: bearer
      bearerFormat: JWT